+++ warden/app/ante/ante.go
@@ -5,9 +5,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
 	authante "github.com/cosmos/cosmos-sdk/x/auth/ante"
-	"github.com/cosmos/gogoproto/proto"
-
-	schedv1beta1 "github.com/warden-protocol/wardenprotocol/api/warden/sched/v1beta1"
 )
 
 // NewAnteHandler returns an ante handler responsible for attempting to route an
@@ -25,15 +22,15 @@ func NewAnteHandler(options HandlerOptions) sdk.AnteHandler {
 			opts := txWithExtensions.GetExtensionOptions()
 			if len(opts) > 0 {
 				switch typeURL := opts[0].GetTypeUrl(); typeURL {
-				case "/" + proto.MessageName(&schedv1beta1.ExtensionOptionsCallbacks{}):
-					// x/sched evm transaction it's a system transaction bypasses normal checks
-					anteHandler = newSchedAnteHandler(options)
+				// case "/" + proto.MessageName(&schedv1beta1.ExtensionOptionsCallbacks{}):
+				// 	// x/sched evm transaction it's a system transaction bypasses normal checks
+				// 	anteHandler = newSchedAnteHandler(ctx, options)
 				case "/cosmos.evm.vm.v1.ExtensionOptionsEthereumTx":
 					// handle as *evmtypes.MsgEthereumTx
-					anteHandler = newMonoEVMAnteHandler(options)
+					anteHandler = newMonoEVMAnteHandler(ctx, options)
 				case "/cosmos.evm.types.v1.ExtensionOptionDynamicFeeTx":
 					// cosmos-sdk tx with dynamic fee extension
-					anteHandler = newCosmosAnteHandler(options)
+					anteHandler = newCosmosAnteHandler(ctx, options)
 				default:
 					return ctx, errorsmod.Wrapf(
 						errortypes.ErrUnknownExtensionOptions,
@@ -48,7 +45,7 @@ func NewAnteHandler(options HandlerOptions) sdk.AnteHandler {
 		// handle as totally normal Cosmos SDK tx
 		switch tx.(type) {
 		case sdk.Tx:
-			anteHandler = newCosmosAnteHandler(options)
+			anteHandler = newCosmosAnteHandler(ctx, options)
 		default:
 			return ctx, errorsmod.Wrapf(errortypes.ErrUnknownRequest, "invalid transaction type: %T", tx)
 		}